+++ src/library/components/row-selector/select-all.tsx
@@ -2,12 +2,14 @@ import React, { useContext } from 'react';
 import { ColumnContext, useTableSelector } from '../table/contexts';
 import { RowSelectorCheckbox } from './checkbox';
 import isEqual from 'lodash/isEqual';
+import { TableDataContext } from '../table/data-provider';
 
 interface AllRowsSelectorProps<T> {
-  data?: T[]
+  
 }
 
 export const AllRowsSelector = function AllRowsSelector<T>({ ...props }: AllRowsSelectorProps<T>) {
+  const { data } = useContext(TableDataContext);
   const {
     canSelectRow,
     setAllSelected,
@@ -18,7 +20,7 @@ export const AllRowsSelector = function AllRowsSelector<T>({ ...props }: AllRows
   
   const Checkbox = components?.RowCheckbox ?? RowSelectorCheckbox;
 
-  const rowData = props.data ?? [];
+  const rowData = data ?? [];
   const filteredRowData = rowData.filter(rw => (typeof canSelectRow !== 'function' || canSelectRow(rw)));
   const selectedKeys = Object.keys(selectedRows);
 
@@ -32,7 +34,7 @@ export const AllRowsSelector = function AllRowsSelector<T>({ ...props }: AllRows
         indeterminate={isIndeterminite}
         onChange={() => {
           if (!filteredRowData.length) return;
-          setAllSelected(!!filteredRowData.length && selectedKeys.length !== filteredRowData.length)
+          setAllSelected(!!filteredRowData.length && selectedKeys.length !== filteredRowData.length, filteredRowData)
         }}
       />
     </div>