+++ src/library/components/table/body.tsx
@@ -9,8 +9,10 @@ import { useArrayDerivedState } from '../../utils/useDerivedState';
 import { getGroupKey, getGroupKeys, GroupKey } from '../../utils/getGroupKey';
 import { TableGroup } from './table-group';
 import { update } from '../../utils/immutable';
+import { TableDataContext } from './data-provider';
 
-export const TableBody = function TableBody<T>({ data, loading, canEditRow, LoadingComponent, ...props }: TableBodyProps<T>) {
+export const TableBody = function TableBody<T>({ canEditRow, LoadingComponent, ...props }: TableBodyProps<T>) {
+  const { data, loading } = useContext(TableDataContext);
   const { actualColumns: columns, groupBy } = useContext(ColumnContext);
   const tbodyRef = useRef<HTMLTableSectionElement>(null);