+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+### Added
+- Refresh button and configurable refetch handler.
+- New method of supplying data to the datatable - the overrideable DataProvider.
 
 ## [1.7.0]
 ### Added